P's CR fixes
